+++ DailyJournal/client/src/components/Entry/EntryProvider.js
