+++ DailyJournal/client/src/components/Mood/MoodProvider.js
